fixed file seperation accordingly
