+++ game_logics.go
@@ -155,3 +155,49 @@ func onTigerPress(boardX, boardY int) {
 	}
 }
 
+func isValidMove(from, to [2]int) bool {
+	// Ensure destination is within range
+	if to[0] < 0 || to[1] < 0 || to[0] >= 5 || to[1] >= 5 {
+			return false
+	}
+
+	// Ensure the destination is connected to the source
+	validMoves, exists := validConnections[from]
+	if !exists {
+			return false
+	}
+	isConnected := false
+	for _, conn := range validMoves {
+			if conn == to {
+					isConnected = true
+					break
+			}
+	}
+	if !isConnected {
+			return false
+	}
+
+	// Check adjacency or valid jump (already handled)
+	dx := to[0] - from[0]
+	dy := to[1] - from[1]
+
+	// Single step (adjacency)
+	if (abs(dx) == 1 && dy == 0) ||
+			(dx == 0 && abs(dy) == 1) ||
+			(abs(dx) == 1 && abs(dy) == 1) {
+			return true
+	}
+
+	// Two-step jump with a goat in the middle
+	if (abs(dx) == 2 && dy == 0) ||
+			(dx == 0 && abs(dy) == 2) ||
+			(abs(dx) == 2 && abs(dy) == 2) {
+			midX := (from[0] + to[0]) / 2
+			midY := (from[1] + to[1]) / 2
+			if boardState[midX][midY] == 1 { // Must be a goat in the middle
+					return true
+			}
+	}
+
+	return false
+}
